CBMH-ML/requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
crayons==0.3.0
cryptography==2.8
decorator==4.4.2
Django==3.0.3
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
django-suit==0.2.28
drf-yasg==1.17.1
json5==0.9.0
jsonschema==3.2.0
numpy==1.18.1
openapi-codec==1.3.2
opencv-python==4.2.0.34
pickleshare==0.7.5
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
swagger-spec-validator==2.5.0
Pillow==7.0.0
crayons==0.3.0
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
Jinja2==2.10.3
joblib==0.14.1
numpydoc==0.9.1
pandas==0.25.3
scikit-image==0.15.0
scikit-learn==0.22
scipy==1.3.2
urllib3==1.25.7
coreapi==2.3.3
coreschema==0.0.4
crayons==0.3.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.2